// Card_Game_Engine/ClientApp/src/app/shared/models/enums/color.enum.ts
export enum ColorEnum {
  Blue = '#2e99fa',
  Gray = '#d5d4d4',
  Green = '#a2d752',
  Yellow = '#ffd753',
  Red = '#ff9696',
  Orange = '#ffb347',

  Pink = '#FFB3BA',
  Peach = '#FFDFBA',
  LightYellow = '#FFFFBA',
  Mint = '#baffc0',
  Sky = '#BAE1FF',
  Lavender = '#E6CCFF',
  PastelPurple = '#F0D9FF',
  SoftGreen = '#D9EAD3',
  LightGold = '#FFF2CC',
  BabyBlue = '#CFE2F3',

  DarkBlue = '#002e5b',
  DarkGray = '#3e3e3e',
  DarkGreen = '#2e4a00',
  DarkYellow = '#614a01',
  DarkRed = '#6a0000',
  DarkOrange = '#cc5500',

  DarkPink = '#800000',
  DarkPeach = '#7F4F24',
  DarkLightYellow = '#665c00',
  DarkMint = '#006644',
  DarkSky = '#003366',
  DarkLavender = '#4B0082',
  DarkPastelPurple = '#5D3A9B',
  DarkSoftGreen = '#274E13',
  DarkLightGold = '#7F6000',
  DarkBabyBlue = '#134F5C',
}

export const playersColors: { background: ColorEnum; color: ColorEnum }[] = [
  {background: ColorEnum.Peach, color: ColorEnum.DarkPeach},
  {background: ColorEnum.Sky, color: ColorEnum.DarkSky},
  {background: ColorEnum.PastelPurple, color: ColorEnum.DarkPastelPurple},
  {background: ColorEnum.Mint, color: ColorEnum.DarkMint},
  {background: ColorEnum.LightYellow, color: ColorEnum.DarkLightYellow},
  {background: ColorEnum.BabyBlue, color: ColorEnum.DarkBabyBlue},
  {background: ColorEnum.LightGold, color: ColorEnum.DarkLightGold},
  {background: ColorEnum.SoftGreen, color: ColorEnum.DarkSoftGreen},
  {background: ColorEnum.Lavender, color: ColorEnum.DarkLavender},
  {background: ColorEnum.Pink, color: ColorEnum.DarkPink},
];